Scraper/pyproject.toml
[tool.poetry]
name = "scraper"
version = "0.1.0"
description = ""
authors = ["fwrrong <fwrrong@umich.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
temporalio = "^1.5.0"
selenium = "^4.17.2"
bson = "^0.5.10"
beautifulsoup4 = "^4.12.3"
overrides = "^7.7.0"
pymongo = "^4.6.1"
boto3 = "^1.34.37"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
